Decode the inbox on initial get
This takes the inbox, which is the content field of the initial get request in
base64 encoding, and decodes it. Added a lot of tests for error
handling. I'm a little unhappy with the error handling logic because I
ran into issues with the borrow checker and couldn't use the map_err
functions the way I wanted to.