Add AppDelegate category to detect login
On iOS when the user logs in the browser redirects to
paytonrules.Capture:// - which is not a valid URL but does open this app
because it is registered with a custom URL scheme.

The app opens, and we detect it in our new AppDelegate which is the same
one that Godot created but with a new category (OpenURL) to add the
callback.

This required adding header files from Godot to the project, so they
could be referenced.