Refactor provider tests for clarity
Instead of using launch to send the port and state, make an explicit
function that does it in the tests (so you have a clear 'act' step).
This added a ton of extra Arc/Mutex/RefCell stuff which is a bummer.

Rename new to new_with_state on the MockTokenReceiver.