Write Webserver trait to take a callback
I had to comment out a lot of tests here and fight a lot with the borrow
checker, the compiler and even lifetimes. Webserver now has a launch
function which takes the callback. To get that tested required some
work.

Provide will become a single threaded function when this is completed,
with work moved into the webserver for startup/shutdown.